+++ R/tar_test.R
@@ -26,7 +26,7 @@ tar_test <- function(label, code) {
     tar_assert_package("testthat")
     code <- substitute(code)
     expr <- substitute(
-      tar_dir(testthat::test_that(label, code)),
+      targets::tar_dir(testthat::test_that(label, code)),
       env = list(label = label, code = code)
     )
     withr::local_envvar(c(TAR_ASK = "false", TAR_TEST = "true"))