+++ NEWS.md
@@ -1,8 +1,13 @@
 # targets 0.8.0.9000
 
+## New features
+
+* Add arguments `strict` and `silent` to allow `tar_load()` and `tar_load_raw()` to bypass targets that cannot be loaded.
+
 ## Enhancements
 
 * Improve `tidyselect` docs in `tar_make()` (#640, @dewoller).
+* Use namespaced call to `tar_dir()` in `tar_test()` (#642, @billdenney).
 
 # targets 0.8.0